# backend/webandtech/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/products/{id}:
    get:
      tags:
      - products-rest-controller
      summary: Get a products by its id
      operationId: getProduct
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Product not found
          content: {}
        "200":
          description: Found the Product
          content:
            application/json: {}
    put:
      tags:
      - products-rest-controller
      summary: Modify a Product
      operationId: replaceProduct
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product_Detailed'
        required: true
      responses:
        "404":
          description: Product not found
          content: {}
        "201":
          description: Successful Product modification
          content:
            application/json: {}
        "406":
          description: Not Acceptable Post title exists
    delete:
      tags:
      - products-rest-controller
      summary: Delete a Product
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Product not found
          content: {}
        "201":
          description: Successful Product delete
          content:
            application/json: {}
...
  /api/products/{id}/ratings:
    get:
      tags:
      - products-rest-controller
      summary: Ratings
      operationId: getProductRatings
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        description: page
        required: false
        schema:
          type: string
      responses:
        "404":
          description: ratings not found
          content: {}
        "200":
          description: Found the ratings Product
          content:
            application/json: {}
        "204":
          description: ratings not found
    post:
      tags:
      - products-rest-controller
      summary: Ratings
      operationId: setProductRatings
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ratingtext:
                  type: string
                  description: Text of rating
                iduser:
                  type: integer
                  description: Id user
                  format: int32
        required: true
      responses:
        "404":
          description: ratings not found
          content: {}
        "200":
          description: Found the ratings Product
          content:
            application/json: {}
        "204":
          description: ratings not found
  /api/products/{id}/image3:
    get:
      tags:
      - products-rest-controller
      summary: Get a Image 3 Product by its id
      operationId: getImage3
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Product not found
          content: {}
        "200":
          description: Found the Image Product
          content:
            application/json: {}
        "204":
          description: Image not found
    post:
      tags:
      - products-rest-controller
      summary: Create a Image 3 Product by its id
      operationId: uploadImage3
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  description: Image 3 Product
                  format: binary
      responses:
        "404":
          description: Product not found
          content: {}
        "201":
          description: Create the Image Product
          content:
            application/json: {}
        "204":
          description: Image not found
  /api/products/{id}/image2:
    get:
      tags:
      - products-rest-controller
      summary: Get a Image 2 Product by its id
      operationId: getImage2
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Product not found
          content: {}
        "200":
          description: Found the Image Product
          content:
            application/json: {}
        "204":
          description: Image not found
    post:
      tags:
      - products-rest-controller
      summary: Create a Image 2 Product by its id
      operationId: uploadImage2
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  description: Image 2 Product
                  format: binary
      responses:
        "404":
          description: Product not found
          content: {}
        "201":
          description: Create the Image Product
          content:
            application/json: {}
        "204":
          description: Image not found
  /api/products/{id}/image1:
    get:
      tags:
      - products-rest-controller
      summary: Get a Image 1 Product by its id
      operationId: getImage1
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Product not found
          content: {}
        "200":
          description: Found the Image Product
          content:
            application/json: {}
        "204":
          description: Image not found
    post:
      tags:
      - products-rest-controller
      summary: Create a Image 1 Product by its id
      operationId: uploadImage1
      parameters:
      - name: id
        in: path
        description: id of Product to be searched
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  description: Image 1 Product
                  format: binary
      responses:
        "404":
          description: Product not found
          content: {}
        "201":
          description: Create the Image Product
          content:
            application/json: {}
        "204":
          description: Image not found
  /api/products/:
    get:
      tags:
      - products-rest-controller
      summary: Get New Six Products
      operationId: getAllProducts
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Found the Products
          content:
            application/json: {}
    post:
      tags:
      - products-rest-controller
      summary: Create a Products
      operationId: registerProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product_Detailed'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable Post title exists
        "201":
          description: Successful Products creation
          content:
            application/json: {}
...
  /api/products/phones:
    get:
      tags:
      - products-rest-controller
      summary: Get a phone
      operationId: getPhonesPage
      parameters:
      - name: page
        in: query
        description: page
        required: false
        schema:
          type: string
      - name: typeProduct
        in: query
        description: typeProduct
        required: false
        schema:
          type: string
      responses:
        "404":
          description: Product not found
          content: {}
        "200":
          description: Found the Product
          content:
            application/json: {}
  /api/products/peripherals:
    get:
      tags:
      - products-rest-controller
      summary: Get a peripheral
      operationId: getPeripheralsPage
      parameters:
      - name: page
        in: query
        description: page
        required: false
        schema:
          type: string
      - name: typeProduct
        in: query
        description: typeProduct
        required: false
        schema:
          type: string
      responses:
        "404":
          description: Product not found
          content: {}
        "200":
          description: Found the Product
          content:
            application/json: {}
  /api/products/components:
    get:
      tags:
      - products-rest-controller
      summary: Get a component
      operationId: getComponentsPage
      parameters:
      - name: page
        in: query
        description: page
        required: false
        schema:
          type: string
      - name: typeProduct
        in: query
        description: typeProduct
        required: false
        schema:
          type: string
      responses:
        "404":
          description: Product not found
          content: {}
        "200":
          description: Found the Product
          content:
            application/json: {}
...
components:
  schemas:
...
    Product_Detailed:
      type: object
      properties:
        idproduct:
          type: integer
          format: int32
        nameproduct:
          type: string
        price:
          type: number
          format: float
        description:
          type: string
        ratingProducto:
          type: number
          format: float
        productType:
          type: string
        productcategory:
          type: string
        productbrand:
          type: string
        img1:
          type: boolean
        img2:
          type: boolean
        img3:
          type: boolean
      description: Object Type Product
...